Remove interactions
