+++ lib/renderer/dom/index.js
@@ -3,7 +3,6 @@
  * Dependencies.
  */
 
-var Interactions = require('./interactions');
 var virtualize = require('../../virtualize');
 var shallowCopy = require('shallow-copy');
 var each = require('component-each');
@@ -15,17 +14,45 @@ var walk = require('dom-walk');
 var zip = require('array-zip');
 var isDom = require('is-dom');
 var uid = require('get-uid');
-var tree = virtualize.tree;
+var throttle = require('per-frame');
+var keypath = require('object-path');
 
 /**
- * Prevent calling setState in these lifecycle states
- *
- * @type {Object}
+ * All of the events can bind to
  */
 
-var preventSetState = {
-  beforeUpdate: "You can't call setState in the beforeUpdate hook. Use the propsChanged hook instead.",
-  render: "You can't call setState in the render hook. This method must remain pure."
+var events = {
+  onBlur: 'blur',
+  onChange: 'change',
+  onClick: 'click',
+  onContextMenu: 'contextmenu',
+  onCopy: 'copy',
+  onCut: 'cut',
+  onDoubleClick: 'dblclick',
+  onDrag: 'drag',
+  onDragEnd: 'dragend',
+  onDragEnter: 'dragenter',
+  onDragExit: 'dragexit',
+  onDragLeave: 'dragleave',
+  onDragOver: 'dragover',
+  onDragStart: 'dragstart',
+  onDrop: 'drop',
+  onFocus: 'focus',
+  onInput: 'input',
+  onKeyDown: 'keydown',
+  onKeyUp: 'keyup',
+  onMouseDown: 'mousedown',
+  onMouseMove: 'mousemove',
+  onMouseOut: 'mouseout',
+  onMouseOver: 'mouseover',
+  onMouseUp: 'mouseup',
+  onPaste: 'paste',
+  onScroll: 'scroll',
+  onSubmit: 'submit',
+  onTouchCancel: 'touchcancel',
+  onTouchEnd: 'touchend',
+  onTouchMove: 'touchmove',
+  onTouchStart: 'touchstart'
 };
 
 /**
@@ -42,6 +69,9 @@ module.exports = dom;
  */
 
 function dom(app) {
+  var frameId;
+  var isRendering;
+
   setup();
 
   /**
@@ -49,9 +79,10 @@ function dom(app) {
    */
 
   function setup() {
-    app.events = new Interactions(document.body);
+    addNativeEventListeners();
     app.entities = {};
     app.pools = {};
+    app.handlers = {};
     app.on('unmount component', onunmount);
     app.on('mount component', onmount);
     app.on('update component', onupdate);
@@ -62,6 +93,9 @@ function dom(app) {
    */
 
   function teardown() {
+    removeNativeEventListeners();
+    delete app.pools;
+    delete app.handlers;
     app.off('unmount component', onunmount);
     app.off('mount component', onmount);
     app.off('update component', update);
@@ -126,8 +160,6 @@ function dom(app) {
     entity.virtualElement = virtualElement;
     entity.nativeElement = nativeElement;
 
-    updateEvents(entity);
-
     // callback after mounting.
     trigger('afterMount', entity, [
       nativeElement,
@@ -153,7 +185,7 @@ function dom(app) {
       entity.props,
       entity.state
     ]);
-    removeEvents(entity);
+    removeAllEvents(entity);
     delete app.entities[entity.id];
   }
 
@@ -167,10 +199,8 @@ function dom(app) {
 
   function renderEntity(entity) {
     var component = entity.component;
-    entity.lifecycle = 'render';
     var result = component.render(entity.props, entity.state, send);
     if (!result) result = virtualize.node('noscript');
-    entity.lifecycle = null;
     return result;
 
     // Whenever setState or setProps is called, we mark the entity
@@ -200,8 +230,6 @@ function dom(app) {
    * @api public
    */
 
-  var frameId;
-  var isRendering;
   function render() {
 
     // If this is called synchronously we need to
@@ -259,7 +287,6 @@ function dom(app) {
     // apply new virtual tree to native dom.
     patch(entity, currentTree, nextTree, el);
     entity.virtualElement = nextTree;
-    updateEvents(entity);
     updateChildren(id);
     var currentProps = entity.props;
     var currentState = entity.state;
@@ -271,29 +298,6 @@ function dom(app) {
       previousProps,
       previousState
     ]);
-  }
-
-  /**
-   * Updates all the DOM event bindings for an entity.
-   *
-   * @param {Entity} entity
-   */
-
-  function updateEvents(entity) {
-    app.events.unbind(entity.id);
-    var nodes = tree(entity.virtualElement).nodes;
-
-    // TODO: Optimize this by storing the events in the Tree
-    // object on the initial pass instead of looping again.
-    // eg. entity.current.events -> '0.0.1:click': fn
-    each(nodes, function(path, node){
-      if (node.type !== 'element') return;
-      each(node.events, function(eventType, fn){
-        app.events.bind(entity.id, path, eventType, function(e){
-          fn.call(entity.component, e, entity.props, entity.state);
-        });
-      });
-    });
   }
 
   /**
@@ -336,18 +340,7 @@ function dom(app) {
 
     // Empty the pools so that our elements don't
     // stay in memory accidentally.
-    app.pools = {};
     app = null;
-  }
-
-  /**
-   * Unbind all events from an entity
-   *
-   * @param {Entity} entity
-   */
-
-  function removeEvents(entity) {
-    app.events.unbind(entity.id);
   }
 
   /**
@@ -416,7 +409,7 @@ function dom(app) {
     }
 
     // set attributes.
-    for (var name in attributes) setAttribute(el, name, attributes[name]);
+    for (var name in attributes) setAttribute(entityId, path, el, name, attributes[name]);
 
     // store keys on the native element for fast event handling.
     el.__entity__ = entityId;
@@ -518,7 +511,7 @@ function dom(app) {
    * Diff the attributes and add/remove them.
    */
 
-  function diffAttributes(prev, next, el) {
+  function diffAttributes(prev, next, el, entity, path) {
     var nextAttrs = next.attributes;
     var prevAttrs = prev.attributes;
 
@@ -526,14 +519,14 @@ function dom(app) {
     for (var name in nextAttrs) {
       var value = nextAttrs[name];
       if (!prevAttrs[name] || prevAttrs[name] !== value) {
-        setAttribute(el, name, value);
+        setAttribute(entity.id, path, el, name, value);
       }
     }
 
     // remove old attrs
     for (var oldName in prevAttrs) {
       if (!nextAttrs[oldName]) {
-        el.removeAttribute(oldName);
+        removeAttribute(entity.id, path, el, oldName);
       }
     }
   }
@@ -560,7 +553,7 @@ function dom(app) {
     // different node, so swap them. If the root node of the component has changed it's
     // type we need to update this to point to this new element
     if (next.tagName !== prev.tagName) return replaceElement(entity, path, el, next);
-    diffAttributes(prev, next, el);
+    diffAttributes(prev, next, el, entity, path);
     diffChildren(path, entity, prev, next, el);
   }
 
@@ -688,7 +681,11 @@ function dom(app) {
    * @param {String} value
    */
 
-  function setAttribute(el, name, value) {
+  function setAttribute(entityId, path, el, name, value) {
+    if (events[name] != null) {
+      addEvent(entityId, path, events[name], value);
+      return;
+    }
     switch (name) {
       case 'value':
         el.value = value;
@@ -700,6 +697,22 @@ function dom(app) {
         el.setAttribute(name, value);
         break;
     }
+  }
+
+  /**
+   * Remove an attribute, performing additional transformations
+   * dependning on the attribute name
+   *
+   * @param {HTMLElement} el
+   * @param {String} name
+   */
+
+  function removeAttribute(entityId, path, el, name) {
+    if (events[name] != null) {
+      removeEvent(entityId, path, events[name]);
+      return;
+    }
+    el.removeAttribute(name);
   }
 
   /**
@@ -790,12 +803,10 @@ function dom(app) {
    */
 
   function trigger(name, entity, args) {
-    entity.lifecycle = name;
     if (typeof entity.component[name] === 'function') {
       args.push(send); // last arg is `send`
       entity.component[name].apply(null, args);
     }
-    entity.lifecycle = null;
 
     /**
      * Function for hook to set state if desired.
@@ -813,22 +824,9 @@ function dom(app) {
    */
 
   function updateEntityState(entity, nextState) {
-    checkSetState(entity.lifecycle);
     entity.pendingState = assign(entity.pendingState, nextState);
     entity.dirty = true;
     invalidate();
-  }
-
-  /**
-   * Determine whether it is possible to set state during a
-   * lifecycle method.
-   *
-   * @param {String} lifecycle
-   */
-
-  function checkSetState(lifecycle) {
-    var message = preventSetState[lifecycle];
-    if (message) throw new Error(message);
   }
 
   /**
@@ -929,6 +927,87 @@ function dom(app) {
     });
   }
 
+  /**
+   * Add all of the DOM event listeners
+   */
+
+  function addNativeEventListeners() {
+    Object.keys(events).forEach(function(attr){
+      var eventType = events[attr];
+      document.body.addEventListener(eventType, handleEvent, true);
+    })
+  }
+
+  /**
+   * Add all of the DOM event listeners
+   */
+
+  function removeNativeEventListeners() {
+    Object.keys(events).forEach(function(attr){
+      var eventType = events[attr];
+      document.body.removeEventListener(eventType, handleEvent, true);
+    })
+  }
+
+  /**
+   * Handle an event that has occured within the container
+   *
+   * @param {Event} event
+   */
+
+  function handleEvent(event) {
+    var target = event.target;
+    var entityId = target.__entity__;
+    var eventType = event.type;
+
+    // Walk up the DOM tree and see if there is a handler
+    // for this event type higher up.
+    while (target && target.__entity__ === entityId) {
+      var fn = keypath.get(app.handlers, [entityId, target.__path__, eventType]);
+      if (fn) {
+        event.delegateTarget = target;
+        fn(event);
+        break;
+      }
+      target = target.parentNode;
+    }
+  }
+
+  /**
+   * Bind events for an element, and all it's rendered child elements.
+   *
+   * @param {String} path
+   * @param {String} event
+   * @param {Function} fn
+   */
+
+  function addEvent(entityId, path, eventType, fn) {
+    keypath.set(app.handlers, [entityId, path, eventType], throttle(function(e){
+      var entity = app.entities[entityId];
+      fn.call(null, e, entity.props, entity.state);
+    }));
+  }
+
+  /**
+   * Unbind events for a entityId
+   *
+   * @param {String} entityId
+   */
+
+  function removeEvent(entityId, path, eventType) {
+    keypath.del(app.handlers, [entityId, path, eventType]);
+  }
+
+  /**
+   * Unbind all events from an entity
+   *
+   * @param {Entity} entity
+   */
+
+  function removeAllEvents(entity) {
+    keypath.del(app.handlers, [entity.id]);
+  }
+
   return teardown;
 };
 
@@ -948,7 +1027,6 @@ function Entity(component, props) {
   this.props = props || {};
   this.component = component;
   this.state = this.component.initialState(this.props);
-  this.lifecycle = null;
   this.pendingProps = assign({}, this.props);
   this.pendingState = assign({}, this.state);
   this.dirty = false;
