+++ lib/virtualize/element.js
@@ -7,44 +7,6 @@ var type = require('component-type');
 module.exports = ElementNode;
 
 /**
- * All of the events can bind to
- */
-
-var events = {
-  onBlur: 'blur',
-  onChange: 'change',
-  onClick: 'click',
-  onContextMenu: 'contextmenu',
-  onCopy: 'copy',
-  onCut: 'cut',
-  onDoubleClick: 'dblclick',
-  onDrag: 'drag',
-  onDragEnd: 'dragend',
-  onDragEnter: 'dragenter',
-  onDragExit: 'dragexit',
-  onDragLeave: 'dragleave',
-  onDragOver: 'dragover',
-  onDragStart: 'dragstart',
-  onDrop: 'drop',
-  onFocus: 'focus',
-  onInput: 'input',
-  onKeyDown: 'keydown',
-  onKeyUp: 'keyup',
-  onMouseDown: 'mousedown',
-  onMouseMove: 'mousemove',
-  onMouseOut: 'mouseout',
-  onMouseOver: 'mouseover',
-  onMouseUp: 'mouseup',
-  onPaste: 'paste',
-  onScroll: 'scroll',
-  onSubmit: 'submit',
-  onTouchCancel: 'touchcancel',
-  onTouchEnd: 'touchend',
-  onTouchMove: 'touchmove',
-  onTouchStart: 'touchstart'
-};
-
-/**
  * Initialize a new `ElementNode`.
  *
  * @param {String} tagName
@@ -57,7 +19,6 @@ var events = {
 function ElementNode(tagName, attributes, key, children) {
   this.type = 'element';
   this.attributes = parseAttributes(attributes);
-  this.events = parseEvents(attributes);
   this.tagName = parseTag(tagName, attributes);
   this.children = children || [];
   this.key = key;
