+++ lib/virtualize/index.js
@@ -6,7 +6,6 @@
 var ComponentNode = require('./component');
 var ElementNode = require('./element');
 var TextNode = require('./text');
-var tree = require('./tree');
 var slice = require('sliced');
 var uid = require('get-uid');
 
@@ -15,7 +14,6 @@ var uid = require('get-uid');
  */
 
 exports.node = dom;
-exports.tree = tree;
 
 /**
  * Create virtual DOM trees.
