+++ test/virtualize/index.js
@@ -2,5 +2,4 @@ mocha.setup({ globals: ['hasCert'] });
 
 describe('Virtual DOM', function(){
   require('./node');
-  require('./tree');
 });
